Add eslint, move models to separate file, and remove module flag.
