+++ index.js
@@ -1,78 +1,18 @@
-import express from "express";
-import { render } from "preact-render-to-string";
-import BigSequelize from "sequelize";
-import session from "express-session";
-import bodyParser from "body-parser";
+const express = require('express');
+const { render } = require('preact-render-to-string');
+const session = require('express-session');
+const bodyParser = require('body-parser');
+
+const { User, Topic, Forum, dbSync } = require('./models');
+const { HtmlPage } = require("./template.js");
+const { IndexPage } = require("./pages/index.js");
+const { ForumPage } = require("./pages/forum.js");
+const { TopicPage } = require("./pages/topic.js");
 
 const SESSION_SECRET = "just trust me";
 
-const { Sequelize, DataTypes } = BigSequelize;
-
 const app = express();
 const PORT = process.env.PORT || 8080;
-
-const sequelize = new Sequelize("sqlite:db.sqlite3");
-const User = sequelize.define(
-  "User",
-  {
-    username: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    bio: {
-      type: DataTypes.STRING,
-      allowNull: true,
-    },
-  },
-  {}
-);
-const Forum = sequelize.define(
-  "Forum",
-  {
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-  },
-  {}
-);
-const Topic = sequelize.define(
-  "Topic",
-  {
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-  },
-  {}
-);
-const Message = sequelize.define("Message", {
-  title: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-  body: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-  date: {
-    type: DataTypes.DATE,
-    allowNull: false,
-  },
-});
-
-Forum.hasMany(Topic);
-Topic.belongsTo(Forum);
-Topic.hasMany(Message);
-Message.belongsTo(Topic);
-User.hasMany(Message);
-Message.belongsTo(User);
-Topic.hasOne(User);
-
-import { HtmlPage } from "./template.js";
-import { IndexPage } from "./pages/index.js";
-import { ForumPage } from "./pages/forum.js";
-import { TopicPage } from "./pages/topic.js";
 
 const renderPage = (title, page) => {
   return HtmlPage({ title, content: render(page) });
@@ -97,7 +37,6 @@ app.get("/forum/:forumId", async (req, res) => {
     )
   );
 });
-
 app.get("/topic/:topicId", async (req, res) => {
   const topic = await Topic.findByPk(req.params.topicId);
   const messages = await topic.getMessages({ include: User });
@@ -115,40 +54,6 @@ app.get("/dbsync", async (req, res) => {
   res.end("aw hell yea");
 });
 
-async function dbSync() {
-  await sequelize.sync({ force: true });
-
-  const potch = await User.create({
-    username: "potch",
-    bio: "dababy enthusiast",
-  });
-  const osmose = await User.create({
-    username: "osmose",
-    bio: "final fantasy-type beat",
-  });
-  const forum = await Forum.create({ name: "Title Screen" });
-  const topic = await Topic.create({
-    name: "the first topic",
-    // forumId: forum.id,
-  });
-  await topic.setForum(forum);
-
-  const message1 = await Message.create({
-    title: "Whatup fam",
-    body: "hows it hanging",
-    date: Date.now(),
-  });
-  await message1.setUser(osmose);
-  await message1.setTopic(topic);
-  const message2 = await Message.create({
-    title: "Re: Re: Whatup fam",
-    body: "nm how bout u",
-    date: Date.now(),
-  });
-  await message2.setUser(potch);
-  await message2.setTopic(topic);
-}
-
 const start = async () => {
   app.use("/wp-content", express.static("wp-content"));
 
@@ -164,7 +69,7 @@ const start = async () => {
       resave: false,
       saveUninitialized: true,
       cookie: {
-        maxAge: 60 * 525_602,
+        maxAge: 60 * 525602,
       },
     })
   );
