+++ scripts/dbSync.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+const { dbSync } = require('../models');
+
+dbSync();
